+++ console/src/dataSources/services/bigquery/index.tsx
@@ -156,7 +156,7 @@ export const supportedFeatures: DeepRequired<SupportedFeaturesType> = {
         hostSource: true,
         referenceSource: true,
       },
-      remoteRelationships: false,
+      remoteRelationships: true,
     },
     permissions: {
       enabled: true,